% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{BC.AD}
\alias{BC.AD}
\title{Converts year number (epoch) to calendar year}
\usage{
BC.AD(year)
}
\arguments{
\item{year}{year number}
}
\description{
Converts year number (epoch) to calendar year
}
\examples{
BC.AD(100)
BC.AD(0)
BC.AD(-1)
BC.AD(-99)
BC.AD(-100)
}
