% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellations-data.R
\docType{data}
\name{constellations}
\alias{constellations}
\title{Constellations}
\format{
A data frame with 300 rows and 2 variables:
 \describe{
   \item{set}{Object identifier used for ephemeris}
   \item{name1}{Full name of constellation}
   \item{name2}{Shorter name for constellation}
   \item{name3}{Alternative name for constellation}
   \item{name4}{Alternative name for constellation}
   \item{HIP}{String with HIP catalogue number of stars to join together.
   This should be in the same format that Stellarium uses for its sky cultures.}
   }
}
\usage{
data(constellations)
}
\description{
Constellations database. Currently only IAU / Modern western constellations
included, but can be expanded to include those of other societies.
}
\keyword{datasets}
