% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{createHor}
\alias{createHor}
\title{Create skyscapeR.horizon object from Az/Alt data}
\usage{
createHor(az, alt, lat, lon, name)
}
\arguments{
\item{az}{Array of azimuth values}

\item{alt}{Array of altitude values}

\item{lat}{Latitude of site}

\item{lon}{Longitude of site}

\item{name}{Name of site}
}
\description{
This function creates a skyscapR.horizon object from measurements of
azimuth and altitude.
}
\examples{
# Create a skyscapeR.horizon from 5 measurements:
az <- c(0,90,180,270,360)
alt <- c(0,5,5,0,0)
hor <- createHor(az, alt, 40.1, -8, 'Test')
plot(hor)
}
\seealso{
\code{\link{plot.skyscapeR.horizon}}
}
