% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{curvigram}
\alias{curvigram}
\title{Computes declination curvigram}
\usage{
curvigram(dec, sd = 2, norm = T)
}
\arguments{
\item{dec}{Array of declination values}

\item{sd}{Either an array of declination values (same length as dec),
or a single value to be applied to all declinations. Defaults to 2.}

\item{norm}{Boolean specifying whether the resulting curvigram
should be normalized to unity. Defaults to \emph{TRUE}.}
}
\description{
This function computes the curvigram of declinations,
using provided measurement uncertainty and a Gaussian
kernel.
}
\examples{
# Curvigram of Ruggles' Recumbent Stone Circle data:
data(RugglesRSC)
curv <- curvigram(RugglesRSC$Dec, 2)
plotCurv(curv)
}
