% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{distSolarRange}
\alias{distSolarRange}
\title{Declination distribution of the Sun throughout the year}
\usage{
distSolarRange(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the distribution.
Defaults to present year as given by Sys.Date()}
}
\description{
This function returns the declination distribution of
the the Sun throughout the year. It can be used for significance testing.
}
\examples{
\dontrun{
aux <- distSolarRange(-4000)
plot(aux$dec, aux$density, type='l')
}
}
\seealso{
\code{\link{sigTest}}, \code{\link{distRandom}}, \code{\link{distSummerFM}}, \code{\link{distLunarRange}}
}
