% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{distStars}
\alias{distStars}
\title{Declination distribution of the Brightest Stars}
\usage{
distStars(brightest = 20, precision = 2, year = cur.year)
}
\arguments{
\item{brightest}{(Optional) This should be the number of
bright stars to be considered. Defaults to 20.}

\item{precision}{(Optional) This should be the intended
precision for each declination. Defaults to 2º.}

\item{year}{Year for which to calculate the distribution
Defaults to present year as given by Sys.Date().}
}
\description{
This function returns the declination distribution of
the brightest stars. It can be used for significance testing.
}
\examples{
aux <- distStars(10, 1, -4000)
plot(aux$dec, aux$density, type='l')
}
\seealso{
\code{\link{sigTest}}, \code{\link{distRandom}}, \code{\link{distSolarRange}}, \code{\link{distLunarRange}}
}
