% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{distSummerFM}
\alias{distSummerFM}
\title{Declination distribution of the Summer Full Moon}
\usage{
distSummerFM(min.phase = 0.99, min.sundec = 20, year = cur.year)
}
\arguments{
\item{min.phase}{(Optional) This should be the minimum
lunar phase (i.e. percentage illumination) for the moon
 to be considered full. The value should range between
 0 (dark moon) and 1 (full moon). Defaults to 0.99.}

\item{min.sundec}{(Optional) This should be the minimum
solar declination for the moon to be considered a
\emph{summmer} full moon. Defaults to 20 degrees,
corresponding to a month before or after june solstice.}

\item{year}{Year for which to calculate the distribution.
Defaults to present year as given by Sys.Date()}
}
\description{
This function returns the declination distribution of
the Summer Full Moon. It can be used for significance testing.
}
\examples{
\dontrun{
aux <- distSummerFM(.99, 20, -4000)
plot(aux$dec, aux$density, type='l')
}
}
\seealso{
\code{\link{sigTest}}, \code{\link{distRandom}}, \code{\link{distSolarRange}}, \code{\link{distLunarRange}}
}
