% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{downloadHWT}
\alias{downloadHWT}
\title{Download horizon data from \emph{HeyWhatsThat}}
\usage{
downloadHWT(HWTID)
}
\arguments{
\item{HWTID}{This is the 8 character ID attributed by
\emph{HeyWhatsThat.com}}
}
\description{
This function downloads previously created horizon data
from \emph{HeyWhatsThat}, given its ID, and saves it as
a \emph{skyscapeR.horizon} object.
}
\examples{
\dontrun{
# Retrieve horizon data for \href{https://www.heywhatsthat.com/?view=HIFVTBGK}{Liverpool Cathedral}:
hor <- downloadHWT('HIFVTBGK')
}
}
\references{
\href{http://heywhatsthat.com/}{HeyWhatsThat.com}
}
\seealso{
\code{\link{createHWT}}
}
