% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{dS}
\alias{dS}
\title{Topodecentric declination of December Solstice for a given year}
\usage{
dS(year = cur.year, parallax = 0.00224, topoalt = 0, geolat = 50)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}

\item{parallax}{Solar parallax value. Defaults to 0.00224 the average parallax value. If set to 0: geocentric declination.}

\item{topoalt}{altitude value. Defaults to 0 (not that much difference if appalt is used)}

\item{geolat}{Geographic lattiude value. Defaults to 50}
}
\description{
This function calculates the Topodecentric declination of the sun
at December Solstice for a given year, based upon
obliquity estimation.
}
\examples{
# December Solstice declination for year 3999 BC:
dS(-4000)
}
\seealso{
\code{\link{obliquity}}, \code{\link{jS}}, \code{\link{eq}}, \code{\link{zenith}}, \code{\link{antizenith}}
}
