% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{eq2horFS}
\alias{eq2horFS}
\title{Fixed eq2hor function from astrolibR package}
\usage{
eq2horFS(ra, dec, jd, lat = 43.0783, lon = -89.865, ws = F, obsname,
  b1950, precess_ = TRUE, nutate_ = TRUE, refract_ = TRUE,
  aberration_ = TRUE, altitude = 0, ...)
}
\description{
Fixed eq2hor function from astrolibR package
}
