% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{exportHor}
\alias{exportHor}
\title{Exports a skyscapeR.horizon object into Stellarium format}
\usage{
exportHor(hor, name, author = "skyscapeR", description, ground_col, hor_col)
}
\arguments{
\item{hor}{Horizon data in skyscapeR.horizon format.}

\item{name}{Horizon name to be displayed in Stellarium, if different from one in skyscapeR.horizon object.}

\item{author}{Author, to be included in landscape.ini file. Optional.}

\item{description}{Description, to be included in landscape.ini file. Optional.}

\item{ground_col}{Colour of ground. Defaults to Stellarium's default.}

\item{hor_col}{Colour of horizon line. Defaults to Stellarium's default.}
}
\description{
This function exports any skyscapR.horizon object into the landscape
format of Stellarium, ready to be imported.
}
\examples{
# Downloads horizon data from HeyWhatsThat and exports it into Stellarium:
hor <- download.HWT('HIFVTBGK')
exportHor(hor, name='Test', description='Test horizon export to Stellarium')
}
\references{
\href{http://www.stellarium.org/}{Stellarium: a free open source planetarium}
}
\seealso{
\code{\link{createHor}}, \code{\link{download.HWT}}, \code{\link{plot.skyscapeR.horizon}}
}
