% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analysis.R
\name{findTargets}
\alias{findTargets}
\title{Function to find celestial targets within declination and time ranges}
\usage{
findTargets(decrange, timerange, nstars = 100, loc = FALSE)
}
\arguments{
\item{decrange}{Range of declination to consider.}

\item{timerange}{Temporal range to consider}

\item{nstars}{(Optional) NUmber of stars to consider i ndescending order of apparent
magnitude. Defaults to 100.}

\item{loc}{Location, either a \emph{skyscapeR.object} or a vector
containing the latitude, longitude and elevation of location, in this order. Defaults
to FALSE, thus checking only geocentric declinations.}
}
\description{
Function to find celestial targets within declination and time ranges
}
\examples{
findTargets(c(-35,-32), c(-2500,-1750))
}
