% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analysis.R
\name{histogram}
\alias{histogram}
\title{Computes declination histograms}
\usage{
histogram(dec, unc = 2, norm = F, cut = 4, range)
}
\arguments{
\item{dec}{Array of declination values.}

\item{unc}{(Optional) Uncertainty in declination values,
which will be used as bin size. Must be a single value.}

\item{norm}{(Optional) Boolean specifying whether the resulting curvigram
should be normalized to unity. Defaults to \emph{FALSE}.}

\item{cut}{(Optional) Number of uncertainties beyond the extremes
of the data at which to trim the curvigram. Defaults to 4. See \code{\link{density}}.}

\item{range}{(Optional) As an alternative to \emph{cut} you can
stipulate the range of declination values to output as an array of two values.
See \emph{from, to} in \code{\link{density}}.}
}
\description{
This function computes the histogram of declinations,
using provided measurement uncertainty and a bin size.
}
\examples{
# Histogram of Ruggles' Recumbent Stone Circle data:
data(RugglesRSC)
hist <- histogram(RugglesRSC$Dec, 2)
plot(hist)
}
\seealso{
\code{\link{curvigram}}
}
