% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon.R
\name{hor2max.dec}
\alias{hor2max.dec}
\title{Retrieves maximum declination for a given horizon}
\usage{
hor2max.dec(hor, alt = 0, range = 25)
}
\arguments{
\item{hor}{Either a \emph{skyscapeR.horizon} object or the latitude of the place.}

\item{alt}{(Optional) Altitude of the horizon at northernmost point. Only necessary if no
\emph{skyscapeR.horizon} object is provided. Default is 0.}

\item{range}{(Optional) Range of azimuths to check for maximum declination. Default is 25
degrees either side of North.}
}
\description{
Retrieves maximum declination for a given horizon
}
\examples{
hor <- downloadHWT('HIFVTBGK')
hor2max.dec(hor)
}
