% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hpdi}
\alias{hpdi}
\title{Returns the high-density region of a probability distribution}
\usage{
hpdi(x, mass = 0.954)
}
\arguments{
\item{x}{A \emph{skyscapeR.spd} or \emph{skyscapeR.pdf} object.}

\item{mass}{(Optional) Probability mass of the region. Default is 0.954.}
}
\description{
Returns the high-density region of a probability distribution
}
