% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{long.date}
\alias{long.date}
\title{Converts day-month in 'MM-DD' format to a more readable format}
\usage{
long.date(date)
}
\arguments{
\item{date}{date in 'MM-DD' format}
}
\description{
Converts day-month in 'MM-DD' format to a more readable format
}
\examples{
long.date('01-01')
long.date('08-23')
}
