% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malta_horizons-data.R
\docType{data}
\name{malta_horizons}
\alias{malta_horizons}
\title{Maltese Temples horizon profiles}
\format{
A large list of 32 skyscapeR.horizon elements.
}
\usage{
data(malta_horizons)
}
\description{
Horizon profiles from Tore Lomsdalen's fieldwork on the prehistoric Maltese Temples.
}
\examples{
data(malta_horizons)
plot(malta_horizons[[1]])
plot(malta_horizons[[10]])
}
\references{
Lomsdalen, T. (2021). \emph{Viewscapes and Cosmology in the Prehistoric Temples of Malta}. PhD Dissertation - University of Malta.

Silva, F. and Lomsdalen, T. (2025). ‘No easy way from the earth to the stars’: a new statistical approach to the orientation of the Maltese Temples. \emph{Archaeological and Anthropological Sciences} <doi:10.1007/s12520-025-02208-4>
}
\keyword{datasets}
