% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{modelTest}
\alias{modelTest}
\title{Significance test against a null model}
\usage{
modelTest(
  pdf,
  model,
  nsims = 1000,
  conf = 0.95,
  tails = 2,
  normalise = F,
  ncores = parallel::detectCores() - 1,
  save.sim = F,
  verbose = T
)
}
\arguments{
\item{pdf}{A \emph{skyscapeR.pdf} object created with either \code{\link{az.pdf}} or \code{\link{coordtrans}}}

\item{model}{A \emph{data.frame} or list of \emph{data.frames} with equal number to the empirical observations.
See example for structure}

\item{nsims}{(Optional) Number of simulations to run. Default is 1000.}

\item{conf}{(Optional) Confidence envelope (in percentage) of the null model to calculate. Default is .95}

\item{tails}{(Optional) Whether to calculate 1-tailed p-value (greater than) or 2-tailed p-value (smaller than or greater than).
Default is 2.}

\item{normalise}{(Optional) Boolean to control whether to normalize SPDs. Default is FALSE}

\item{ncores}{(Optional) Number of CPU cores to use. Default is the number of available cores minus 1.}

\item{verbose}{(Optional) Boolean to control whether or not to display text. Default is TRUE.}

\item{.res}{(Optional) Resolution with which to output probability distribution(s). Default is 0.1 degrees.}
}
\description{
Significance test against a null model
}
