% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{nMjLX}
\alias{nMjLX}
\title{Topodecentric declination of northern major Lunar Extreme for a given year}
\usage{
nMjLX(year = cur.year, parallax = 0.952)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}

\item{parallax}{Lunar parallax value. Defaults to 0.952 the average parallax value. If set to 0: geocentric declination.}
}
\description{
This function calculates the Topodecentric declination of the northern
major Lunar Extreme for a given year, by simple addition
of obliquity with maximum lunar inclination.
}
\examples{
# Northern major Lunar Extreme Topodecentric declination for year 2499 BC:
nMjLX(-2500)
}
\seealso{
\code{\link{nmnLX}}, \code{\link{smnLX}}, \code{\link{sMjLX}}
}
