% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{nmnLX}
\alias{nmnLX}
\title{Topodecentric declination of northern minor Lunar Extreme for a given year}
\usage{
nmnLX(year = cur.year, parallax = 0.952, topoalt = 0)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}

\item{parallax}{Lunar parallax value. Defaults to 0.952 the average parallax value. If set to 0: geocentric declination.}

\item{topoalt}{altitude value. Defaults to 0 (not that much difference if appalt is used)}
}
\description{
This function calculates the Topodecentric declination of the northern
minor Lunar Extreme for a given year, by simple subtraction
of obliquity with maximum lunar inclination and perturbation.
}
\examples{
# Northern minor Lunar Extreme Topodecentric declination for year 2499 BC:
nmnLX(-2500)
}
\seealso{
\code{\link{smnLX}}, \code{\link{nMjLX}}, \code{\link{sMjLX}}
}
