% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{object}
\alias{object}
\title{Creates a _skyscapeR.object_ for plotting of celestial objects at given epoch}
\usage{
object(names, epoch, col = "red", lty = 1, lwd = 1)
}
\arguments{
\item{epoch}{The year or year range (as an array) one is interested in.}

\item{col}{(Optional) The colour for plotting, and differentiating these objects.
Defaults to red for all objects.}

\item{lty}{(Optional) Line type (see \code{\link{par}}) used for differentiation.
Only activated for single year epochs.}

\item{lwd}{(Optional) Line width (see \code{\link{par}}) used for differentiation.
Only activated for single year epochs.}

\item{name}{The name(s) of the celestial object(s) or event(s) of interest.
These can be one of the following soli-lunar events: _jS_, _dS_, _eq_, _nmnLX_, _nMjLX_,
_smnLX_, _sMjLX_, or the name of any star in the database. As shorthand, the names
_sun_ and _moon_ can be used to represent all the above solar and lunar events,
respectively. Alternatively, a custom declination value can be used.}
}
\description{
This function creates an object containing all the necessary information to
plot celestial objects/events unto the many plotting functions of _skyscapeR_
package.
}
\examples{
# Create a object with solar targets for epoch range 4000-2000 BC:
tt <- object('sun', c(-4000,-2000))

# Create an object with a few stars for same epoch:
tt <- object(c('Sirius', 'Betelgeuse', 'Antares'), c(-4000,-2000), col=c('white', 'red', 'orange'))

# Create an object with solstices and a custom declination value:
tt <- object(c('dS','jS', -13), c(-4000,-2000))
}
