% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{obliquity}
\alias{obliquity}
\title{Computes obliquity of the ecliptic}
\usage{
obliquity(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the obliquity.
Defaults to present year as given by Sys.Date()}
}
\description{
This function calculates the true obliquity (so incl. nutation) for a given epoch. It is a
wrapper for function \code{\link[swephR]{swe_calc_ut}} of package \emph{swephR}.
}
\examples{
#' # Obliquity for year 3999 BC:
obliquity(-4000)
}
\seealso{
\code{\link[swephR]{swe_calc_ut}}
}
