% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{orbit}
\alias{orbit}
\title{Calculate visible path of celestial object at given location}
\usage{
orbit(dec, lat, lon)
}
\arguments{
\item{dec}{Declination of object.}

\item{lat}{Latitude of location}

\item{lon}{Longitude of location}
}
\description{
This function calculates the visibile path of a celestial
object from any location on earth. It outputs a 'skyscapeR.orbit'
object, which includes AZ and ALT information.
}
\examples{
# Visible path of sun on June Solstice on year 3999 BC from London:
sun.dec <- jS(-4000)
london.lat <- 51.5074 #N
london.lon <- -0.1278 #W
path <- orbit(sun.dec, london.lat, london.lon)
plot(path$az, path$alt, ylim=c(0,90), type='l', xlab='AZ', ylab='ALT', col='red', lwd=2)
}
