% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{orplot}
\alias{orplot}
\title{Polar plot of orientations (azimuths)}
\usage{
orplot(az, ...)
}
\arguments{
\item{az}{Array of azimuths. Values outside the [0, 360] range will be ignored.}

\item{...}{Any other parameters to be passed unto \code{\link[plotrix]{polar.plot}}}
}
\description{
This function creates a polar plot of azimuthal data. It is a wrapper for
\code{\link[plotrix]{polar.plot}}
}
\examples{
# Plot some azimuth data:
az <- c(120, 100, 93, 97, 88, 115, 112, 67)
orplot(az)
}
\seealso{
\code{\link[plotrix]{polar.plot}}
}
