% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{palaeo.star}
\alias{palaeo.star}
\title{Time shift object of class 'skyscapeR.star'}
\usage{
palaeo.star(star, year = cur.year)
}
\arguments{
\item{star}{Object created using \code{\link{star}}}

\item{year}{Year for which to calculate the coordinates.
Defaults to present year as given by Sys.Date()}
}
\description{
This function calculates the coordinates (RA and DEC)
of a star for any time in the past. It uses functions
\code{\link[astrolibR]{precess}}, \code{\link[astrolibR]{co_nutate}}
and \code{\link[astrolibR]{co_aberration}}.
}
\examples{
# Time shift data on Sirius to 2500 BC:
Sirius <- star('Sirius')
Sirius.2500BC <- palaeo.star(Sirius, -2501)
}
\seealso{
\code{\link[astrolibR]{precess}}, \code{\link[astrolibR]{co_nutate}},
\code{\link[astrolibR]{co_aberration}}
}
