% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{planet.extremes}
\alias{planet.extremes}
\title{Extreme declinations for five visible planets}
\usage{
planet.extremes(obj, epoch, loc = NULL)
}
\arguments{
\item{obj}{The name of the planet.}

\item{epoch}{Epoch for which to do calculations.}

\item{loc}{(Optional) This can be either a \emph{skyscapeR.horizon} object, or a vector with the
latitude, longitude and elevation of the site, in this order.}
}
\description{
This function calculates the maximum and minimum declinations of the five visible planets
for a given epoch. This is done by calculating the declination for one entire orbital period centred
around the epoch.
}
\examples{
planet.extremes('Jupiter', 2023)

planet.extremes('Venus', -4000)
}
