% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetarium.R
\name{planisphere}
\alias{planisphere}
\title{Create a planisphere view of the sky for a given epoch and location}
\usage{
planisphere(
  epoch,
  lat,
  max.mag = 6,
  show.ecliptic = TRUE,
  col.ecliptic = "orange",
  show.equator = TRUE,
  col.equator = "red",
  star.col = "black"
)
}
\arguments{
\item{epoch}{Numerical.The epoch (year) for which the planisphere is calculated.}

\item{lat}{Numerical.Latitude of the observer in degrees.}

\item{max.mag}{Numerical.Maximum magnitude of stars to be displayed (default: 6).}

\item{show.ecliptic}{Optional.Whether to display the ecliptic (TRUE or FALSE).}

\item{col.ecliptic}{Character.Colour of the ecliptic line (default: "orange").}

\item{show.equator}{Optional.Whether to display the celestial equator (True or False).}

\item{col.equator}{Character.Colour of the celestial equator (default: "red").}

\item{star.col}{Character.Colour of the stars (default: "black").}
}
\value{
A polar plot of the visible sky, showing stars, the celestial equator,
        and the ecliptic for the specified epoch and location.
}
\description{
This function generates a planisphere, which is a circular star chart
that represents the visible sky for a given epoch and observer's latitude.
}
\examples{
planisphere(2023, lat=50, max.mag=6, show.ecliptic=TRUE, show.equator=TRUE)
}
