% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.hist}
\alias{plot.skyscapeR.hist}
\title{Plot a histogram}
\usage{
\method{plot}{skyscapeR.hist}(hh, obj, col = "blue", xrange,
  legend = F)
}
\arguments{
\item{hh}{Object of \emph{skyscapeR.hist} format, created using \code{\link{histogram}}.}

\item{obj}{(Optional) A \emph{skyscapeR.object} object created with \code{\link{sky.objects}}
for displaying the declination of celestial objects.}

\item{col}{(Optional) Colour to plot the histogram in. Defaults to blue.}

\item{xrange}{Array of two values restricting the horizontal range of the plot.}

\item{legend}{(Optional) Show legend. Defaults to FALSE.}
}
\description{
This function creates a plot of a histogram
}
\examples{
# Plot the histogram of Recumbent Stone Circles:
data(RugglesRSC)
hist <- histogram(RugglesRSC$Dec, unc=2)
plot(hist, xrange=c(-40,0))

# Redo the plot to include lunar extreme declinations:
LEx <- sky.objects('moon', -2000, col='red', lty=2)
plot(hist, obj=LEx, xrange=c(-40,0))
}
\seealso{
\code{\link{histogram}}, \code{\link{curvigram}}, \code{\link{sky.objects}}
}
