% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.horizon}
\alias{plot.skyscapeR.horizon}
\title{Plot horizon data}
\usage{
\method{plot}{skyscapeR.horizon}(
  hor,
  obj,
  data,
  show.unc = F,
  show.ground = T,
  show.axes = F,
  xrange,
  yrange,
  legend = F
)
}
\arguments{
\item{hor}{Object of \emph{skyscapeR.horizon} format.}

\item{obj}{(Optional) A \emph{skyscapeR.object} object created with \code{\link{sky.objects}}
for displaying the paths of celestial objects.}

\item{data}{(Optional) A data.frame object with columns \emph{True.Azimuth} and}

\item{show.unc}{(Optional) Boolean that controls whether to display uncertainty in altitude.
Default is \emph{FALSE}.}

\item{show.ground}{(Optional) Boolean that controls whether to dispaly the ground. Default is \emph{TRUE}.}

\item{show.axes}{(Optional) Boolean that controls whether to display azimuth values on horizontal
axis and altitude on vertical axis. Default is \emph{FALSE}.}

\item{xrange}{Range of azimuth axis. Defaults to c(0, 360).}

\item{yrange}{Range of altitude axis. Defaults to c(-10, 45).}

\item{legend}{(Optional) Show legend. Defaults to FALSE.}
}
\description{
This function creates a plot of horizon data.
}
\examples{
# Plot a horizon retrieved from HeyWhatsThat:
hor <- downloadHWT('HIFVTBGK')
plot(hor)

# Add the paths of the solstices and equinoxes sun in the year 1999 BC:
tt <- sky.objects('sun', -2000, 'blue')
plot(hor, obj=tt)
}
\seealso{
\code{\link{downloadHWT}}, \code{\link{sky.objects}}
}
