% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.orbit}
\alias{plot.skyscapeR.orbit}
\title{Plot visible path of celestial object on horizon}
\usage{
\method{plot}{skyscapeR.orbit}(orbit, col, ...)
}
\arguments{
\item{orbit}{Object of \emph{skyscapeR.orbit} format.}

\item{col}{String with colour to plot path in. Defaults to red.}
}
\description{
This function adds the visible path of a celestial object to a horizon plot.
}
\examples{
hor <- download.HWT('HIFVTBGK')
plot(hor)

path <- orbit(dS(-2500),hor)
plot(path, col='blue')
}
\seealso{
\code{\link{plot.skyscapeR.horizon}}, \code{\link{orbit}}
}
