% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.pdf}
\alias{plot.skyscapeR.pdf}
\title{Plot orientation probability distributions}
\usage{
\method{plot}{skyscapeR.pdf}(
  x,
  index,
  hdr = 0.954,
  show.az = T,
  xlim,
  col = MESS::col.alpha("blue", 0.5),
  ...
)
}
\arguments{
\item{x}{A \emph{skyscapeR.pdf} object created with either \code{\link{az.pdf}} or \code{\link{coordtrans}}}

\item{index}{(Optional) A value to indicate which distribution to plot, when only one is desired.}

\item{hdr}{(Optional) High density region to highlight. Defaults to 0.954}

\item{show.az}{(Optional) Whether to show the azimuth and transformation curve (horizon) when
displaying declination distributions. Defaults to TRUE.}

\item{xlim}{(Optional) Range of x-axis to plot.}

\item{col}{(Optional) Color of high density region to highlight. Defaults to blue with 0.5 alpha.}

\item{...}{Additional arguments to be passed to \emph{plot}.}
}
\description{
Plot orientation probability distributions
}
