% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.spd}
\alias{plot.skyscapeR.spd}
\title{Plot orientation summed probability density}
\usage{
\method{plot}{skyscapeR.spd}(x, xlim, ylim, title = NULL, col = "blue", shading = T, ...)
}
\arguments{
\item{x}{A \emph{skyscapeR.spd} object created with \code{\link{spd}}}

\item{xlim}{(Optional) Range of x-axis to plot.}

\item{ylim}{(Optional) Range of y-axis to plot.}

\item{title}{(Optional) Title to add to the plot.}

\item{col}{(Optional) Color of summed probability density. Defaults to blue with 0.5 alpha.}

\item{shading}{(Optional) Boolean to control whether to color the entire distribution. Defaults to TRUE}

\item{...}{Additional arguments to be passed to \emph{plot}.}
}
\description{
Plot orientation summed probability density
}
