% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.starphases}
\alias{plot.skyscapeR.starphases}
\title{Plot stellar phase and seasonality}
\usage{
\method{plot}{skyscapeR.starphases}(
  x,
  y,
  show.labels = TRUE,
  show.legend = TRUE,
  show.grid = FALSE,
  sort.by,
  pal,
  ...
)
}
\arguments{
\item{x}{Object of \emph{skyscapeR.starphases} format.}

\item{y}{(Optional) A second object of \emph{skyscapeR.starphases} format or
a list of \emph{skyscapeR.starphases} objects for plotting (see Examples).}

\item{show.labels}{(Optional) Logical. If `TRUE`, labels will be shown in the plot. Default is `FALSE`.}

\item{show.legend}{(Optional) Only used when plotting seasonality of multiple
stars (see example below). Default is TRUE.}

\item{show.grid}{(Optional) Boolean to control whether to show a grid of months.
Default is FALSE}

\item{sort.by}{(Optional) Character. Defines how data should be sorted before plotting. Options: `"time"`, `"azimuth"`.}

\item{pal}{(Optional) Colour palette for displaying phases/seasons.}

\item{...}{Additional arguments to be passed to \emph{plot}.}
}
\description{
This function creates a plot of stellar seasonality and phases/events.
}
\examples{
# Plot the seasonality of Aldebaran for 3999 BCE:
s1 <- star.phases('Aldebaran',-4000, c(35,-8, 200))
plot(s1)
# Plot seasonality of several stars
s2 <- star.phases('Alnilam',-4000, c(35,-8, 200))
s3 <- star.phases('Elnath',-4000, c(35,-8, 200))
plot(s1, s2)
plot(s1, list(s2,s3))
plot(s1, show.legend=FALSE, show.grid=TRUE)
}
\seealso{
\code{\link{star.phases}}
}
