% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAz}
\alias{plotAz}
\title{Polar plot of orientations (azimuths)}
\usage{
plotAz(az, obj, loc, col = "blue", lwd = 1, lty = 1, len = 1,
  legend = F)
}
\arguments{
\item{az}{Array of azimuths or data frame with column named \emph{True.Azimuth}. Values
outside the [0, 360] range will be ignored.}

\item{obj}{(Optional) A \emph{skyscapeR.object} object created with \code{\link{sky.objects}}
for displaying the azimuths of celestial objects. Beware that this assumes a single
location (given by parameter loc) and a flat horizon of zero degrees.}

\item{loc}{(Optional) This can be either a vector with the latitude and longitude of the
location, or a \emph{skyscapeR.horizon} object. Only necessary for plotting potential
celestial targets.}

\item{col}{(Optional) Single colour or colour pallete to use for plotting measurements.}

\item{lwd}{(Optional) Line width to plot measurements. Defaults to 1.}

\item{lty}{(Optional) Line type to plot measurements. Defaults to 1.}

\item{len}{(Optional) Length of line to plot measurements.}

\item{legend}{(Optional) Show legend. Defaults to FALSE.}
}
\description{
This function creates a polar plot of azimuthal data using \emph{plotly}.
}
\examples{
# Plot some azimuth data:
az <- c(120, 100, 93, 97, 88, 115, 112, 67)
plotAz(az)

# To visualise this data against the common solar and lunar targets:
tt <- sky.objects(c('sun','moon'), epoch=-2000, lty=1, lwd=1.5, col=c('black','red'))
plotAz(az, obj=tt, loc=c(35,-8), legend=T)

# To visualise the same data with shorter lines
plotAz(az, obj=tt, loc=c(35,-8), len=0.2)

# To display only celestial objects
plotAz(az=NULL, obj=tt, loc=c(35,-8), legend=T)
}
