% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.skyscapeR.sigTest}
\alias{print.skyscapeR.sigTest}
\title{Prints significance test results}
\usage{
\method{print}{skyscapeR.sigTest}(x, ...)
}
\arguments{
\item{x}{Object of \emph{skyscapeR.sigTest} format.}

\item{...}{Additional arguments to be passed to \emph{print}.}
}
\description{
This function prints the results of \code{\link{randomTest}}.
}
\seealso{
\code{\link{randomTest}}
}
