% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{randomTest}
\alias{randomTest}
\title{Significance test against the null hypothesis of random orientation}
\usage{
randomTest(
  pdf,
  .res = 0.1,
  nsims = 1000,
  conf = 0.95,
  tails = 2,
  normalise = F,
  ncores = parallel::detectCores() - 1,
  verbose = T,
  hor,
  refraction,
  atm,
  temp
)
}
\arguments{
\item{pdf}{A \emph{skyscapeR.pdf} object created with either \code{\link{az.pdf}} or \code{\link{coordtrans}}}

\item{.res}{(Optional) Resolution with which to output probability distribution(s). Default is 0.1 degrees.}

\item{nsims}{(Optional) Number of simulations to run. Default is 1000.}

\item{conf}{(Optional) Confidence envelope (in percentage) of the null model to calculate. Default is .95}

\item{tails}{(Optional) Whether to calculate 1-tailed p-value (greater than) or 2-tailed p-value (smaller than or greater than).
Default is 2.}

\item{normalise}{(Optional) Boolean to control whether to normalize SPDs. Default is FALSE}

\item{ncores}{(Optional) Number of CPU cores to use. Default is the number of available cores minus 1.}

\item{verbose}{(Optional) Boolean to control whether or not to display text. Default is TRUE.}

\item{hor}{(Optional) A single \emph{skyscapeR.horizon} object created with \code{\link{createHor}} or \code{\link{downloadHWT}}.
This and the following arguments are only needed if \emph{pdf} contains only azimuths. If provided, the
code will assume that all azimuthal measurements are for the same site with horizon given in \emph{hor} and
that the user want to run the significance test in declination. Coordinate transformation will therefore be done automatically.}

\item{refraction}{(Optional) Whether atmospheric refraction is to be taken into account.
If not given the value set by \code{\link{skyscapeR.vars}} will be used instead.}

\item{atm}{(Optional) Atmospheric pressure for refraction calculation.
If not given the value set by \code{\link{skyscapeR.vars}} will be used instead.}

\item{temp}{(Optional) Atmospheric temperature for refraction calculation.
If not given the value set by \code{\link{skyscapeR.vars}} will be used instead.}
}
\description{
Significance test against the null hypothesis of random orientation
}
\examples{
\dontrun{
# significance test for azimuth
Az <- az.pdf(az=c(87,93,90,110), unc=3)
st1 <- randomTest(Az, nsims=1000)
plot(st1)

# significance test for declination
hor <- createHor(az=c(0,360), alt=c(0,0), loc=c(35,-8,25)) # flat horizon with 0 degrees of altitude
Dec <- coordtrans(Az, hor)
st2 <- randomTest(Dec, nsims=1000)
plot(st2)
}
}
\references{
Silva, F (2020) A probabilistic framework and significance test for the analysis of structural orientations
 in skyscape archaeology \emph{Journal of Archaeological Science} 118, 105138. <doi:10.1016/j.jas.2020.105138>
}
