% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{riseset}
\alias{riseset}
\title{Computes the rising and setting azimuth, declination and time of a Solar System object
for a given location and day}
\usage{
riseset(body = "sun", date, cal = "G", timezone = "GMT", loc,
  refraction = T, atm = 1013.25, temp = 15)
}
\arguments{
\item{body}{(Optional) String containing name of the solar system body of interest. Can be
any of the planets (inc. Pluto), the Moon, the Sun or the Ecliptic. Defaults to 'sun'.}

\item{date}{Either a string containing the date in the format "YYYY-MM-DD"
(see \code{\link{timestring}}), or a numeric containing the julian date (see \code{\link{time2jd}}).}

\item{cal}{(Optional) Calendar used in parameter \emph{date}. G for gregorian and J for julian. Defaults to Gregorian.}

\item{timezone}{(Optional) Timezone for output of rising and setting time either as a known acronym
(eg. "GMT", "CET") or a string with continent followed by country capital (eg. "Europe/London"). See
\link{timezones} for details. Defaults to GMT.}

\item{loc}{Location, either a \emph{skyscapeR.object} or a vector
containing the latitude and longitude of location, in this order.}

\item{refraction}{(Optional) Boolean for whether or not atmospheric refraction should be taken into account.
Defaults to _TRUE_.}

\item{atm}{(Optional) Atmospheric pressure (in mbar). Only needed if \emph{refraction} is set to _TRUE_. Default is 1013.25 mbar.}

\item{temp}{(Optional) Atmospheric temperature (in Celsius). Only needed if \emph{refraction} is set to _TRUE_. Default is 15 degrees.}
}
\description{
Computes the rising and setting azimuth, declination and time of a Solar System object
for a given location and day
}
\examples{
# Rising and setting of the sun on june solstice 2018, from the location of London
riseset('sun', '2018-06-21', loc=c(51.5, 0.11))

# Rising ans setting of the moon on june solstice 2018, using a horizon profile
hor <- downloadHWT('HIFVTBGK') # Liverpool cathedral
riseset('moon', '2018-06-21', loc=hor)
}
\seealso{
\code{\link[swephR]{swe_calc_ut}}, \code{\link[swephR]{swe_azalt}}
}
