% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{riseset.year}
\alias{riseset.year}
\title{Computes the daily rising and setting azimuth, declination and time of a Solar System object
for a given location (and horizon) for a full year}
\usage{
riseset.year(body = "sun", year, cal = "G", timezone = "GMT", loc,
  refraction = T, atm = 1013.25, temp = 15,
  ncores = parallel::detectCores() - 1, verbose = T)
}
\arguments{
\item{body}{(Optional) String containing name of the solar system body of interest. Can be
any of the planets (inc. Pluto), the Moon, the Sun or the Ecliptic. Defaults to 'sun'.}

\item{year}{Year for which to calculate rising and settings.}

\item{cal}{(Optional) Calendar used for output. G for gregorian and J for julian. Defaults to Gregorian.}

\item{timezone}{(Optional) Timezone for output of rising and setting time either as a known acronym
(eg. "GMT", "CET") or a string with continent followed by country capital (eg. "Europe/London"). See
\link{timezones} for details. Defaults to GMT.}

\item{loc}{Location, either a \emph{skyscapeR.object} or a vector
containing the latitude and longitude of location, in this order.}

\item{refraction}{(Optional) Boolean for whether or not atmospheric refraction should be taken into account.
Defaults to _TRUE_.}

\item{atm}{(Optional) Atmospheric pressure (in mbar). Only needed if \emph{refraction} is set to _TRUE_. Default is 1013.25 mbar.}

\item{temp}{(Optional) Atmospheric temperature (in Celsius). Only needed if \emph{refraction} is set to _TRUE_. Default is 15 degrees.}

\item{ncores}{(Optional) Number of processing cores to use for parallelisation. Defaults to the number of
available cores minus 1.}

\item{verbose}{(Optional) Boolean to decide whether or not}
}
\description{
Computes the daily rising and setting azimuth, declination and time of a Solar System object
for a given location (and horizon) for a full year
}
\examples{
# daily rising and setting of the moon in 2006, from the location of London
riseset.year('moon', 2006, loc=c(51.5, 0.11))
}
\seealso{
\code{\link[swephR]{swe_calc_ut}}, \code{\link[swephR]{swe_azalt}}
}
