% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{sMjLX}
\alias{sMjLX}
\title{Declination of southern major Lunar Extreme for a given year}
\usage{
sMjLX(year = cur.year, parallax = 0.85)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}

\item{parallax}{Lunar parallax value. Defaults to 0.85.}
}
\description{
This function calculates the declination of the southern
major Lunar Extreme for a given year, by simple addition
of obliquity with maximum lunar inclination.
}
\examples{
# Southern major Lunar Extreme declination for year 2499 BC:
sMjLX(-2500)
}
\seealso{
\code{\link{nmnLX}}, \code{\link{nMjLX}}, \code{\link{smnLX}}
}
