% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{solar.date}
\alias{solar.date}
\title{Solar Date}
\usage{
solar.date(dec, year, calendar)
}
\arguments{
\item{dec}{Single value or array of declinations.}

\item{year}{Year for which to do calculations.}

\item{calendar}{(Optional) Calendar used for output. G for gregorian and J for julian. Defaults to Gregorian.}
}
\description{
Returns the calendar date when the sun has the same declination as the input declinations.
}
\examples{
solar.date(-23, 2018)
solar.date(-23, 1200, calendar='G')
solar.date(-23, 1200, calendar='J')
}
