% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.names-data.R
\docType{data}
\name{star.names}
\alias{star.names}
\title{Celetial Object Names}
\format{
A data frame with 300 rows and 2 variables:
 \describe{
   \item{identifier}{Object identifier used for ephemeris}
   \item{Western}{Western (IAU / common) name for the object}
   }
}
\usage{
data(star.names)
}
\description{
Celestial object names (stars, supernovae and galaxies) database. Currently only IAU / Western
common names included, but can be expanded to include names from other societies, past or present.
}
\keyword{datasets}
