% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\name{swapStars}
\alias{swapStars}
\title{Swap between swephR and skyscapeR versions of stellar ephemeris}
\usage{
swapStars(source = "skyscapeR")
}
\arguments{
\item{source}{Package name for the source of the stellar ephemeris data. Can be
either \emph{swephR} or \emph{skyscapeR}. Defaults to the latter.}
}
\description{
Swap between swephR and skyscapeR versions of stellar ephemeris
}
