% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ten}
\alias{ten}
\title{Converts degree measurements in deg-min-sec (º ' ") format into decimal-point degree format.}
\usage{
ten(dd, mm = 0, ss = 0)
}
\arguments{
\item{dd}{Degree}

\item{mm}{(Optional) Arcminutes}

\item{ss}{(Optional) Arcseconds}
}
\description{
Converts degree measurements in deg-min-sec (º ' ") format into decimal-point degree format.
}
\examples{
deg <- ten(24, 52, 16)
}
