% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{timestring}
\alias{timestring}
\title{Converts date and time numeric values to a single string}
\usage{
timestring(year, month, day, hour = 12, minute = 0, second = 0)
}
\arguments{
\item{year}{Year}

\item{month}{Month}

\item{day}{Day}

\item{hour}{Hour}

\item{minute}{Minute}

\item{second}{Second}
}
\description{
Converts date and time numeric values to a single string
}
\examples{
timestring(2018, 12, 25, 2, 34)
}
