% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{curvigram}
\alias{curvigram}
\title{(Defunct) Computes declination curvigram}
\usage{
curvigram()
}
\description{
This function no longer works. Please use \code{\link{spd}} of \code{\link{density}} instead.
}
\seealso{
\code{\link{spd}}
}
