% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{dS}
\alias{dS}
\title{Declination of December Solstice for a given year}
\usage{
dS(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}
}
\description{
This function calculates the declination of the sun
at December Solstice for a given year, based upon
obliquity estimation.
}
\examples{
# December Solstice declination for year 3999 BC:
dS(-4000)
}
\seealso{
\code{\link{obliquity}}, \code{\link{jS}}, \code{\link{eq}}, \code{\link{zenith}}, \code{\link{antizenith}}
}
