% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{eq}
\alias{eq}
\title{Declination of sun at the equinoxes}
\usage{
eq(bh = NULL)
}
\arguments{
\item{bh}{\emph{NULL} parameter. Can be left empty.}
}
\description{
This function always returns a value of zero, which is the
declination of the sun on the day of the (astronomical)
equinoxes.
}
\examples{
eq()
}
\seealso{
\code{\link{jS}}, \code{\link{dS}}, \code{\link{zenith}}, \code{\link{antizenith}}
}
