% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{jS}
\alias{jS}
\title{Topodecentric declination of June Solstice for a given year}
\usage{
jS(year = cur.year, parallax = 0.00224)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}

\item{parallax}{Solar parallax value. Defaults to 0.00224 the average parallax value. If set to 0: geocentric declination.}
}
\description{
This function calculates the Topodecentric declination of the sun
at June Solstice for a given year, based upon
obliquity estimation.
}
\examples{
# June Solstice declination for year 3999 BC:
jS(-4000)
}
\seealso{
\code{\link{obliquity}}, \code{\link{dS}}, \code{\link{eq}}, \code{\link{zenith}}, \code{\link{antizenith}}
}
