% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{nh.LunarRange}
\alias{nh.LunarRange}
\title{Declination distribution of the Moon throughout the year
 for null hypothesis significance testing}
\usage{
nh.LunarRange(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the distribution.
Defaults to present year as given by Sys.Date()}
}
\description{
This function returns the declination distribution of
the the Sun throughout the year for use in significance testing.
}
\examples{
\dontrun{
aux <- nh.SolarRange(-4000)
plot(aux$dec, aux$density, type='l')
}
}
\seealso{
\code{\link{sigTest}}, \code{\link{nh.Uniform}}, \code{\link{nh.SummerFM}}
}
