% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{nMjLX}
\alias{nMjLX}
\title{Declination of northern major Lunar Extreme for a given year}
\usage{
nMjLX(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}
}
\description{
This function calculates the declination of the northern
major Lunar Extreme for a given year, by simple addition
of obliquity with maximum lunar inclination.
}
\examples{
# Northern major Lunar Extreme declination for year 2499 BC:
nMjLX(-2500)
}
\seealso{
\code{\link{nmnLX}}, \code{\link{smnLX}}, \code{\link{sMjLX}}
}
