% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{nmnLX}
\alias{nmnLX}
\title{Declination of northern minor Lunar Extreme for a given year}
\usage{
nmnLX(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}
}
\description{
This function calculates the declination of the northern
minor Lunar Extreme for a given year, by simple addition
of obliquity with maximum lunar inclination.
}
\examples{
# Northern minor Lunar Extreme declination for year 2499 BC:
nmnLX(-2500)
}
\seealso{
\code{\link{smnLX}}, \code{\link{nMjLX}}, \code{\link{sMjLX}}
}
