% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.curv}
\alias{plot.skyscapeR.curv}
\title{Plot a curvigram}
\usage{
\method{plot}{skyscapeR.curv}(curv, xlim = NULL, ...)
}
\arguments{
\item{curv}{Object of \emph{skyscapeR.curv} format, created using \code{\link{curvigram}}.}

\item{...}{Any other parameters to be passed unto \code{\link{plot.default}}.}
}
\description{
This function creates a plot of a curvigram
}
\examples{
# Plot some azimuth data:
decs <- data(RugglesRSC)
curv <- curvigram(decs, sd)
plot(curv)
}
\seealso{
\code{\link{curvigram}}
}
