% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.horizon}
\alias{plot.skyscapeR.horizon}
\title{Plot horizon data}
\usage{
\method{plot}{skyscapeR.horizon}(hor, show.az = F, max.alt, az0 = 0,
  zoom = F, ...)
}
\arguments{
\item{hor}{Object of \emph{skyscapeR.horizon} format.}

\item{show.az}{Boolean that controls whether to display azimuth values on horizontal axis.
Defaults to FALSE.}

\item{max.alt}{Maximum altitude to display. Defaults to 45 degrees.}

\item{az0}{Leftmost azimuth of plot. Defaults to 0 degrees, i.e. North at the left.}

\item{zoom}{Boolean that controls whether to provide a zoomed-in view of 100 degrees in
azimuth and 5 degrees of altitude above the horizonline. Defaults to FALSE.}

\item{...}{Any other parameters to be passed unto \code{\link{plot.default}}.}
}
\description{
This function creates a plot of horizon data.
}
\examples{
hor <- download.HWT('HIFVTBGK')
plot(hor)
}
\seealso{
\code{\link{download.HWT}}
}
