% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.sigTest}
\alias{plot.skyscapeR.sigTest}
\title{Plot significance test of orientations}
\usage{
\method{plot}{skyscapeR.sigTest}(
  x,
  xlim,
  title = NULL,
  show.pval = T,
  show.local = F,
  pal = brewer.pal(5, "PRGn")[c(1, 5)],
  ...
)
}
\arguments{
\item{x}{A \emph{skyscapeR.sigTest} object created with \code{\link{randomTest}}}

\item{xlim}{(Optional) Range of x-axis to plot.}

\item{title}{(Optional) Title to add to the plot.}

\item{show.pval}{(Optional) Boolean to control whether to print the global p-value. Default is TRUE.}

\item{show.local}{(Optional) Boolean to control whether to show local regions of significance. Default is FALSE}

\item{pal}{(Optional) Color palette for local regions of significance. Default is brewer.pal(5, 'PRGn')}

\item{...}{Additional arguments to be passed to \emph{plot}.}
}
\description{
Plot significance test of orientations
}
