% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.sigTest}
\alias{plot.skyscapeR.sigTest}
\title{Plot significance test results}
\usage{
\method{plot}{skyscapeR.sigTest}(sig, xlim, show.pval = T,
  show.local = F)
}
\arguments{
\item{sig}{Object of \emph{skyscapeR.sigTest} format, created using \code{\link{sigTest}}.}

\item{xlim}{(Optional) Array of two values restricting the horizontal range of the plot.}

\item{show.pval}{(Optional) Boolean to control whether the global p-value is written.
Default is True.}

\item{show.local}{(Optional) Boolean to control whether the local regions of significance
are plotted. Default is False.}
}
\description{
This function creates a plot of the results of significance test
}
\examples{
\dontrun{
data(RugglesRSC)
curv <- curvigram(RugglesRSC$Dec, sd=2)
sig <- sigTest(curv, null.hyp=nh.Uniform(c(57,2)))

plot(sig, show.local=T)
}
}
\seealso{
\code{\link{sigTest}}
}
