% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.skyscapeR.starphases}
\alias{plot.skyscapeR.starphases}
\title{Plot stellar phase and seasonality}
\usage{
\method{plot}{skyscapeR.starphases}(x, ...)
}
\arguments{
\item{x}{Object of \emph{skyscapeR.starphases} format.}

\item{...}{Additional arguments to be passed to \emph{plot}.}
}
\description{
This function creates a plot of stellar seasonality and phases/events.
}
\examples{
# Plot the seasonality of Aldebaran for 3999 BCE:
\dontrun{
ss <- star.phases('Aldebaran',-4000, c(35,-8, 200))
plot(ss)
}
}
\seealso{
\code{\link{star.phases}}
}
