% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPhases}
\alias{plotPhases}
\title{Plot stellar phase and seasonality}
\usage{
plotPhases(starphase, ...)
}
\arguments{
\item{starphase}{Object of \emph{skyscapeR.starphase} format.}

\item{...}{Any other parameters to be passed unto \code{\link{plot.default}}.}
}
\description{
This function creates a plot of stellar seasonality and phases/events.
}
\examples{
# Plot the seasonality of Aldebaran for 3999 BCE:
\dontrun{
ss <- star.phases('Aldebaran',-4000, c(35,-8))
plotPhases(ss)
}
}
\seealso{
\code{\link{star.phases}}
}
