% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig-test.R
\name{sigTest}
\alias{sigTest}
\title{(Defunct) Perform a null hypothesis significance test of a given curvigram}
\usage{
sigTest()
}
\description{
This function no longer works. Please use \code{\link{randomTest}} instead.
}
\seealso{
\code{\link{randomTest}}
}
