% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyscapeR-package.R
\docType{package}
\name{skyscapeR-package}
\alias{skyscapeR}
\alias{skyscapeR-package}
\title{skyscapeR: Data Analysis and Visualization for Skyscape Archaeology}
\description{
Data reduction, visualization and statistical analysis of measurements of orientation of archaeological structures, following Silva (2020) \doi{10.1016/j.jas.2020.105138}.
}
\author{
\strong{Maintainer}: Silva Fabio \email{fsilva@bournemouth.ac.uk}

Other contributors:
\itemize{
  \item Reijs Victor [contributor]
  \item Gautschy Rita [contributor]
}

}
\keyword{internal}
