% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{smnLX}
\alias{smnLX}
\title{Declination of southern minor Lunar Extreme for a given year}
\usage{
smnLX(year = cur.year)
}
\arguments{
\item{year}{Year for which to calculate the declination.
Defaults to present year as given by \emph{Sys.Date()}.}
}
\description{
This function calculates the declination of the southern
minor Lunar Extreme for a given year, by simple addition
of obliquity with maximum lunar inclination.
}
\examples{
# Southern minor Lunar Extreme declination for year 2499 BC:
smnLX(-2500)
}
\seealso{
\code{\link{nmnLX}}, \code{\link{nMjLX}}, \code{\link{sMjLX}}
}
